게임 시간 단위를 Float 에서 Double 로 변경

+++ src/main/kotlin/game/GameSpeed.kt
@@ -4,11 +4,11 @@ package io.iqpizza.game
  * 게임 속도가 보통을 기준으로 1분 game-loop 가 1350 임을 가정하여,
  * 초당 22.5 로 계산했습니다.
  */
-enum class GameSpeed(val perFrameCount: Float) {
-    SLOWEST(13.5F),
-    SLOW(18F),
-    NORMAL(22.5F),
-    FAST(27F),
-    FASTEST(31.5F)
+enum class GameSpeed(val perFrameCount: Double) {
+    SLOWEST(13.5),
+    SLOW(18.0),
+    NORMAL(22.5),
+    FAST(27.0),
+    FASTEST(31.5)
     ;
 }
